Don't use update_storage_ksdata in tests
Use the storage modules to generate the kickstart data for tests.
